+++ test/key_value_store.js
@@ -499,8 +499,8 @@ describe('KeyValueStore', () => {
         it('should work', async () => {
             process.env[ENV_VARS.LOCAL_STORAGE_DIR] = LOCAL_STORAGE_DIR;
             const defaultStore = await Apify.openKeyValueStore();
-
             // Uses default value.
+            const oldGet = defaultStore.getValue;
             defaultStore.getValue = async key => expect(key).to.be.eql(KEY_VALUE_STORE_KEYS.INPUT);
             await Apify.getInput();
 
@@ -511,6 +511,8 @@ describe('KeyValueStore', () => {
 
             delete process.env[ENV_VARS.LOCAL_STORAGE_DIR];
             delete process.env[ENV_VARS.INPUT_KEY];
+
+            defaultStore.getValue = oldGet;
         });
     });
 });